bruin/zoomcamp/pipeline/assets/ingestion/requirements.txt
# Python dependencies for NYC taxi ingestion asset (trips.py).
# Pin versions for reproducibility.
pandas>=2.0.0
requests>=2.28.0
pyarrow>=14.0.0
python-dateutil>=2.8.0
